[refactor](profile) Remove profile tree printer code (#41766)
Code used to print profile as tree is deleted.
ProfileManager is moved to profile dir.
